refactor(shared): cleaner internal structure
