+++ packages/shared/src/async.ts
@@ -1,0 +1,3 @@
+export function delay(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
