+++ packages/shared/src/auth/index.ts
@@ -1,0 +1,2 @@
+export * from "./auth.input.schema.js";
+export * from "./auth.response.schema.js";
