+++ packages/shared/src/categories/category.query.schema.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { createSortSchema, paginationQuerySchema } from "../query.js";
+import { createSortSchema, paginationQuerySchema } from "../query/index.js";
 
 export const categoryQuerySchema = z
   .object({
