+++ packages/shared/src/categories/index.ts
@@ -1,0 +1,3 @@
+export * from "./category.input.schema.js";
+export * from "./category.query.schema.js";
+export * from "./category.response.schema.js";
