+++ packages/shared/src/comments/index.ts
@@ -1,0 +1,3 @@
+export * from "./comment.input.schema.js";
+export * from "./comment.query.schema.js";
+export * from "./comment.response.schema.js";
