+++ packages/shared/src/common/index.ts
@@ -1,0 +1,2 @@
+export * from "./image.schema.js";
+export * from "./persistence.schema.js";
