+++ packages/shared/src/cuisines/index.ts
@@ -1,0 +1,3 @@
+export * from "./cuisine.input.schema.js";
+export * from "./cuisine.query.schema.js";
+export * from "./cuisine.response.schema.js";
