+++ packages/shared/src/favorites/index.ts
@@ -1,0 +1,1 @@
+export * from "./favorite.query.schema.js";
