+++ packages/shared/src/index.ts
@@ -1,41 +1,14 @@
-// biome-ignore-all assist/source/organizeImports: prevent biome from sorting imports
-
-export * from "./auth/auth.input.schema.js";
-export * from "./auth/auth.response.schema.js";
-
-export * from "./categories/category.input.schema.js";
-export * from "./categories/category.schema.js";
-export * from "./categories/category.response.schema.js";
-
-export * from "./cuisines/cuisine.input.schema.js";
-export * from "./cuisines/cuisine.schema.js";
-export * from "./cuisines/cuisine.response.schema.js";
-
-export * from "./users/user.response.schema.js";
-
-export * from "./comments/comment.input.schema.js";
-export * from "./comments/comment.schema.js";
-export * from "./comments/comment.response.schema.js";
-
-export * from "./recipes/utils.js";
-export * from "./recipes/recipe.primitives.schema.js";
-export * from "./recipes/recipe.input.schema.js";
-export * from "./recipes/recipe.schema.js";
-export * from "./recipes/recipe.response.schema.js";
-
-export * from "./recipes/ingredient.schema.js";
-
-export * from "./favorites/favorite.schema.js";
-
-export * from "./recipe-rating/recipe-rating.input.schema.js";
-
-export * from "./reviews/review.input.schema.js";
-export * from "./reviews/review.response.schema.js";
-export * from "./reviews/review.schema.js";
-
-export * from "./common/image.schema.js";
-export * from "./common/persistence.schema.js";
-
+export * from "./async.js";
+export * from "./auth/index.js";
+export * from "./categories/index.js";
+export * from "./comments/index.js";
+export * from "./common/index.js";
+export * from "./cuisines/index.js";
+export * from "./favorites/index.js";
 export * from "./pagination.js";
-export * from "./utils.js";
-export * from "./query.js";
+export * from "./query/index.js";
+export * from "./recipe-rating/index.js";
+export * from "./recipes/index.js";
+export * from "./reviews/index.js";
+export * from "./type-utils.js";
+export * from "./users/index.js";
