+++ packages/shared/src/query/sort.ts
@@ -49,14 +49,3 @@ export function createSortSchema<const T extends string>(
   const variants = fields.flatMap((f) => [f, `-${f}`] as const);
   return z.enum(variants);
 }
-
-export const searchQuerySchema = z.object({
-  search: z.string().trim().optional(),
-});
-export type SearchQuery = z.infer<typeof searchQuerySchema>;
-
-export const paginationQuerySchema = z.object({
-  page: z.coerce.number().int().min(1).default(1),
-  limit: z.coerce.number().int().min(1).max(100).default(20),
-});
-export type PaginationQuery = z.infer<typeof paginationQuerySchema>;
