+++ packages/shared/src/recipe-rating/index.ts
@@ -1,0 +1,1 @@
+export * from "./recipe-rating.input.schema.js";
