+++ packages/shared/src/recipes/recipe.query.schema.ts
@@ -3,21 +3,12 @@ import {
   createSortSchema,
   paginationQuerySchema,
   searchQuerySchema,
-} from "../query.js";
+} from "../query/index.js";
 import {
   difficultySchema,
   mealTypeSchema,
   minutesSchema,
 } from "./recipe.primitives.schema.js";
-
-export const recipeStatsSchema = z.object({
-  favoritesCount: z.number().int().nonnegative(),
-  commentsCount: z.number().int().nonnegative(),
-  ratingCount: z.number().int().nonnegative(),
-  ratingSum: z.number().int().nonnegative(),
-  averageRating: z.number().nullable(),
-  popularity: z.number().nonnegative(),
-});
 
 export const recipeQuerySchema = z
   .object({
@@ -35,5 +26,4 @@ export const recipeQuerySchema = z
   .extend(paginationQuerySchema.shape)
   .extend(searchQuerySchema.shape);
 
-export type RecipeStats = z.infer<typeof recipeStatsSchema>;
 export type RecipeQuery = z.infer<typeof recipeQuerySchema>;
