+++ packages/shared/src/recipes/recipe.response.schema.ts
@@ -4,7 +4,7 @@ import { persistenceFieldsSchema } from "../common/persistence.schema.js";
 import { cuisineSummarySchema } from "../cuisines/cuisine.response.schema.js";
 import { userSummarySchema } from "../users/user.response.schema.js";
 import { createRecipeInputSchema } from "./recipe.input.schema.js";
-import { recipeStatsSchema } from "./recipe.schema.js";
+import { recipeStatsSchema } from "./recipe.stats.schema.js";
 
 export const recipeComputedSchema = z.object({
   isFavorited: z.boolean(),
