+++ packages/shared/src/reviews/index.ts
@@ -1,0 +1,3 @@
+export * from "./review.input.schema.js";
+export * from "./review.query.schema.js";
+export * from "./review.response.schema.js";
